SAK-40482 LTI copying doesn’t need site.upd
It can be the case that users don’t have the site.upd permission on My Workspace (My Home), this means that when they visit an LTI tool that attempts to patch it’s content it fails because they aren’t allowed to update the LTI tool.

This uses the DAO method which doesn’t do permission checks so it works for people without site.upd permission.