SAKAI-3798 allow instructors to change color of subpage buttons
Instructors now have the following options:
In the "Edit Subpage" dialog, can choose the color for individual subpage buttons.
In the section settings, force all subpage buttons within to be the same color scheme as the section itself
In the Add Layout dialog, choose to force all subpage buttons within to be the same color or not
In the Add Subpage dialog, choose what color the button should be.

SAKAI-3798 allow instructors to change color/style of subpage buttons

moved CSS into namespace for lessonbuilder