SAK-37980 Group visibility issue in Lessons tool (#5911)
When checking to see if an item should be visible there was an "always"
advisor pushed to the top of the security advisor stack before getting
the item's groups. This advisor makes it so that group release information
for a resource is ignored which makes the item stay in the list for the student
even though they don't have access to the resource. This fix pops the advisor
before getting the item groups.