SAK-40520 (#5941)
IMPORTANT NOTES:
For "widgets" that are not tools to be included in this list, they need to have category of widget in tool config.
for widgets that have associated tools, to be added without having the tool installed they need the following property added to their configurations:
<configuration name="require.tool" value="false" type="final" />

A "custom_overview" property is now being saved to the site when Manage Overview is saved for the first time. This property stops synoptics from automatically being added when a tool is added to the site. This DOES carry over to duplicated sites.

This CAN be overwritten in Sites, by editting the custom_overview property to be anything other than true. (Plz for sanity purposes use false).

SAK-40520

revert unnecessary changes