SAK-40817 Assignments regression from SAK-40742 (#6171)
This change cleans up getSubmitterMap and instead of locking the table
it locks the Assignment until it has added all submissions. This change
is a bit smarter since it will only lock the assignment if a submission
is to be added.

(cherry picked from commit ec11ccab1177b4f54e1c6707fd69dfafef5296ea)