SAK-40943 Always include current user memberships
When getting memberships for the current user the list of sites to return the results for were incorrectly getting filtered so that only the entries where the current user was allowed to view the roster in the site. That should have only been happening when the a non-admin user was looking for the memberships of a different user.