SAK-42019 Convert Samigo settings pages to tab layout
This change converts the jquery ui accordion used for Samigo settings to
the jquery ui tabs. The layout is more condusive to working through all
of the settings and provides quicker access and better awareness while
and instructor sets up their assessment. All settings have remained in the
same bucket and the tab names are taken from the accordion titles. There
is a new section "Publish" that contains some information and the save
settings and publish button. This is in a new tab so that users don't accidentally
publish an exam before they are ready. The save and cancel buttons remain at the
bottom of the page.