SAK-41825 SAKAI-2720 Allow users to select a gradebook cateogry in sa… (#7059)
* SAK-41825 SAKAI-2720 Allow users to select a gradebook cateogry in samigo

This re-implements the work previously completed in SAKAI-278 for Sakai 11.
Instructors will be able to select a gradebook category from Samigo
for an assessment's associated gradebook item. Changes were made to the
original implementation to no longer use deprecated classes or methods.

Additionally, this expands upon the original feature to allow for
editing the category setting in published settings. This is done
by pulling in the GB setting everytime published settings is loaded
and then saving a new gradebook item if the category settings have
changed.

* SAK-41825 Use Spring to autowire gradebook service

This changed AssessmentSettingsBean and PublishedAssessmentSettingsBean to use
Spring to autowire the gradebook service, tool manager, and session manager
dependencies.