SAK-34741 SAKAI-2735 SAM-1113 Ability to flag question as Extra Credit (#7060)
* SAK-34741 SAKAI-2735 SAM-1113 Ability to flag question as Extra Credit

This allows instructors to flag questions in their assessment as extra credit. Extra credit questions
will not count towards the total possible points for an assessment. However, any points earned by the
student for answering the extra credit question correctly will be added to their total earned points
for that assessment. Currently, students are able to score over 100 percent on an assessment.