SAK-42221 SAK-34741 Lombok doesn't like boolean fields starting with is (#7195)
Getter for lombok automatically creates boolean accessor as isExtraCredit for
the field isExtraCredit instead of getIsExtraCredit or isIsExtraCredit.
When this is referenced in the template as .isExtraCredit it is not found because
neither getIsExtraCredit nor isIsExtraCredit exists. Changing the name of the
field in ItemBean to extraCredit allows Lombok to create the accessor isExtraCredit
which is then found when using .extraCredit in the template.