SAK-42630 Avoid a possible NPE when catching errors during assessment preview
DeliveryActionListener has a large try/catch block that catches RuntimeException
and then adds an error the error log.

If an exception is caught for an assessment preview, the real error is masked
because the T&Q event logging code attempts to get an assessmentGradingId
which is null (and really there's nothing meaningful to log here since it's a
preview and the event log doesn't log previews).

So we should check for that and throw the real exception earlier before
attempting to update the event log.