SAK-42874 T&Qs remove 'copy from pool' option inside pool menu
Previously this was suposed to be managed by two different selectOneMenu
within the ItemHeadings file, but their selectItems are equal. I also
believe the rendered options were a bit strange, I'd say the only
important check for this is itemauthor.target == 'questionpool' but I'd
like someone else to confirm that, if possible.