SAK-43092 - Fix up POMs for Eclipse compatibility (#7825)
There were a few modules that would compile with Maven directly, but
throw errors when imported as Eclipse projects. The main two issues were
in the Java EE API and a Jackson coordinate change.

There are two minor code changes that I expect would have broken Maven
builds, but apparently did not. One was a test class/package mismatch
in LTI and the other was some missing method implementations for an XML
class in rwiki.