SAK-43201 - Add Integration Test Helper
This addition to the kernel is to provide a convenient base for building
integration tests that run against a fully loaded Spring application
context and ComponentManager.

It builds an uberjar to include all of the dependencies needed to start
the kernel under a test suite and introduces the failsafe plugin to
separate unit and integration tests. Adding the test helper as a single
dependency is sufficient to subclass IntegrationTest for an individual
test or, more likely, a base class for a whole tool and then subsequent
individual tests. The failsafe defaults are used, meaning that test
files starting or ending with "IT" will be run with no additional
configuration.