SAK-43201 - Improve Component Manager to support testing
This commit reworks the relationship between the ComponentManager cover,
SpringCmpMgr, and SakaiApplicationContext. It doesn't change the way
that normal startup happens, but it does refactor and integrate the code
paths some.

There were lots of challenges because of cyclic dependencies causing
synchronization problems. These are not totally gone, but there is a
path forward. Generally speaking, these happen because beans call the
ComponentManager.get methods during their init. It's pretty easy to foul
it up and get transitive or timing issues. The main purpose of this
change is to start the process of inverting the relationship of the
ComponentManager and the Spring context, in order to normalize how beans
are created and wired up.

Key points:

 * There is a new constructor on SpringCmpMgr that allows you to
   create an instance that will initialize itself, rather than the cover
   doing it and needing to load and refresh outside, later.
 * There is a mechanism to register that new instance with the cover
   during its startup so the singleton accesses continue to hit the
   manager before it is refreshed.
 * The IntegrationTestContextLoader uses this new mechanism and gives an
   easy extension point for customizing which configs and components
   are loaded without resorting to files in special paths.
 * The kernel component files are loaded from the classpath, meaning
   that they stay in sync with kernel-impl automatically.
 * The test configuration is also loaded by classpath and defaults to
   reading test.properties from the classpath, meaning the default
   in-memory HSQLDB is automatic.
 * The application context that launches tests sets the Sakai context as
   parent, so all the regular bean config and injection niceties apply
   without doing any explicit test harness initialization.