SAK-43435 Profile2: fix texts next to icons
Font-awesome's own icon class introduced some changes that resulted in a
compression of all the texts next to icons in the Profile tool. Even if
it's not ideal, the best solution seems to be overriding the 'display'
behavior of the component, already on its own css file. The tool adds
classes to components from Java too, so it's hard to trace the number of
cases, but this should solve all of them.