SAK-43435 Profile2: fix texts next to icons (#8028)
Font-awesome's own icon class introduced some changes that resulted in a
compression of all the texts next to icons in the Profile tool. Even if
it's not ideal, the best solution seems to be overriding the 'display'
behavior of the component, already on its own css file. The tool adds
classes to components from Java too, so it's hard to trace the number of
cases, but this should solve all of them.

(cherry picked from commit 87b90f8d805dd412611af4296f6a293d802fe127)