SAK-43472 SWITCH Assignments : use permissions webcomponent (#8074)
This is the first on a series of jiras that will replace the original
permissions screen with the webcomponents one. For that reason, I've
created a common vm file/flow that can be resused by all the velocity
tools. I've renamed/retagged the property files/keys instead of having
to modify the current behavior of the permissions webcomponent.