SAK-43259 Lessons: Error when printing and there's a calendar (#7995)
https://jira.sakaiproject.org/browse/SAK-43259

Providing a working solution for the bug, to get the conversation
started about how to manage it. This problem affects not only the case
described on the jira's title, but also every url access to tools using
moment, and of course the same can happen with other libraries. Adding
library versions on more places is far from ideal, but as long as Sakai
supports this kind of tool access we have to solve the bugs. Even if
that's redesigned we can provide a quick fix first. We can discuss on
this PR about adding a headscripts Java component (a similar feature to
the headscripts js thing on library). DirecttoolHandler or ToolHandler
might be other places where to add it, although the former doesn't cover
every possible call and the latter is executed more times than needed.
It could also be checked against specific tools but I'm a bit against
it.

(cherry picked from commit ecc48ede2081e98e4d50f6580ec3ce574b831752)