SAK-32679 - Handle the handoff from assignments to LTI
This makes sure that LTI links coming from assignments can't be
launched from elsewhere.  It also handles a pretty dynamic set of
situations which would lead to double launching (new window scenario)
that would trigger 403s - instead we force a re-unlock through a
refresh of the "open in a new window" page.