SAK-43856 Moved rubrics webcomponent to webcomponents project (#8317)
https://jira.sakaiproject.org/browse/SAK-43856

Serving the rubrics webcomponents from under a portal url means that
all the requests for the static js files go through the request filter.
If a browser does not forward the cookie in the js request, xlogin
is returned. This fixes that. Rubrics js is now a purely static asset
with no session checks needed.