SAK-39977 remove courier (#8451)
* SAK-44066 Remove courier from portal presence

This removes the functional use of /courier/ URLs from the portal presence tool,
eliminating end-user requests to /courier/... which serve no purpose, because
requesting the presence list itself also serves to establish the user's presence in a site,
and the morpheus js refreshes the presence anyway.

Complete cleanup left for SAK-39977, including removal of scheduleUpdate() from headscripts.js
which is referenced from some other places too (JSF widgets).

Should be mergeable back to 20.x

* SAK-39977 Remove courier code

This touches a lot of code mostly because of the velocity templates. The only
place this seems to be used is in the Admin / Online tool. This changes removes
the auto-refresh functionality.

* SAK-39977 Remove courier code

Further cleanup of unused code

Co-authored-by: Earle Nietzel <earle@longsight.com>