SAK-44373 Fix average gpa calculation in course grade statistics
Average course gpa was calculated incorrectly as a percentage. This
maps the grade labels from the schema to a gpa value when entered in
the format "Grade (gpa)" e.g. "A (4.0)". And then calculates the average
course gpa based on the gpa values taken from the map or n/a if the
labels are not formatted correctly. This used when the grade schema is
set to Grade Points.