SAK-44207 Set hibernate.id.new_generator_mappings back to false after Hibernate upgrade
The Hibernate 5 upgrade switched hibernate.id.new_generator_mappings to true by default
which made it so sequences set in hbm.xml files with class="native" would not get
created when auto.ddl was set to true to create the schema in a blank database. This
switches the property back to false to maintain the behavior before the Hibernate upgrade.