SAK-44890 - Problem running Quartx job "Configurable Job Test"
The fix here seemed to be to disable dependency checking in the new method added in SAK-44246. I tested that SAK-44246 still passes fine with this.
Also changed the logging in Configurable Job Test to info, as I figured if someones actually running this test they'll want to see what it outputs.