SAK-44557 Remove section and column merge link id from href (#8968)
Section and column break item id is needed in javascript to remove the
section or column. This was added to the href of the link but should instead
be added to a data attribute on the link. Links should instead be buttons
and were changed.

(cherry picked from commit 4ca7d3e5f617e95dafde53c47a377795b08cd983)