SAK-44785 Add "direct-upload" endpoint to content entity provider (#8977)
This adds a new endpoint that allows members of a site to upload a file
to resources whether or not they have permission to do so. Those without
permission will have their file sent to a folder called by default
student-uploads. Those with permission to add a resource will be have theirs
sent to instructor-uploads.

This endpoint is diabled by default and can be enabled by setting the sakai
property content.direct.upload.enabled=true. Additionally, the direct-upload
folder names can be overriden in sakai.properties.

This new endpoint integrates with the CKEditor to provide drag and drop embedding
and image upload capabilities.