SAK-45720 Gradebook: category drop settings remain applied
We're applying the changes at gradeRecord level because that's where the
dropped status is checked, plus in case categories are turned on again
it would pick up from where it was in a straightforward way.

Also I had to set the "Object" parameter (just as the method before it),
as Gradebook is in the hibernate package while the service is in the
api.