SAK-33963 Site Info Add automatic and scheduled publish options
This adds two new options to Manage Access for publishing and unpublishing sites.
Specifically, sites can now be set to Auto, Scheduled, or Manual for publishing
and unpublishing. Auto is available for course sites and will use the term dates
along with the existing sakai properties and jobs associated with publishing and
unpublishing sites. Scheduled will allow the site maintainer to pick a date and
time for the site to be published and unpublished. If set to manual the site will
stay in whichever state until the site maintainer updates it.

SAK-33963 Code cleanup use usertimeservice

SAK-33963 Code cleanup and remove jquery in editAccess vm