SAK-46444 Assignments : show closed date error message
View submission page wasn't displaying it either (submission.CloseDate
doesn't even exist). As the same information had to be used on 2 screens
and it was already being processed elsewhere, I refactored two methods
into one and added a tiny piece of logic.
I thought this was the most efficient way to achieve it but of course
feel free to add any recommendations.