SAK-46192 - Refactor Assignments / LTI External Tools For Sakai 21/22 (#9769)
The LTI support for Sakai 21 - worked, but there are a number of places where the Assignments shows information in the UI or is missing information in the UI that makes it hard to use. The basic problem was that all of the handling of scores in Sakai for LTI tools was handled by the Gradebook.  This made much of the Assignments workflow in Sakai pointless or badly broken for LTI Assignments.

The big picture of this change is that for LTI tools placed in Assignments, the grade flow is to/from the Assignments service no longer the grade book service but instead the assignment service and assignments sends grades to the gradebook using its internal logic.  With that change, there are current and emergent LTI grade flow features like activityStatus, gradingStatus, and the submission review launch that Sakai can now support with this change.

This is a dramatic improvement of the LTI integration into Assignments but the good news is that there are no data model changes and this should back port to Sakai 21 nicely.