SAK-46736 - Admin Workspace: Delegated Access - Dark Theme Makes List Unreadable
I moved many of these styles into morpheus-master. There are still some inline styles mostly for the wicket-tree-table component which is deprecated.
These should really all be removed when/if this component is replaced. I did a slight adjustment to make this more useable in darkmode, even though it's got a white background.
Most of the styles for this are internal to the wicket component and not currently overridden.