SAK-46732 UserMembership: put buttons in their own form (#10150)
https://sakaiproject.atlassian.net/browse/SAK-46732

I had to put the command buttons in their own form and remove the
disabled flag. This means the buttons are always enabled which isn't
ideal but causes no harm. If you set disabled in a jsf command button
you can't enable it in JS as the backing component is not updated. It's
always treated as disabled and won't trigger any actions. If I was to do
more work on this I'd literally just rewrite this tool in spring mvc +
thymeleaf.