SAK-46753 - Signup: NPE with email to potential participants
It seems like having itemDisabled here as the default made it so that JSF wouldn't pass any value through to the bean. I don't know if this is something new with JSF2 or not. I'd have to imagine it was working as this was added back in 11. From what I can tell it doesn't need to be disabled by default as there is code client side to enable or disable and this code isn't processed if the checkbox isn't checked.