SAK-46781 Added js log message explaining the 404 error (#10171)
https://sakaiproject.atlassian.net/browse/SAK-46781

This error is displayed by a browser when a 404 is received. Rubrics
makes a get request for the returned evaluation backup copy when saving,
and it's perfectly valid for this to return a 404. The code handles the
404 returned, but the browser also prints the 404 "error" to the
console.

(cherry picked from commit 674386663052ce227f8f9f5f1b1c3002896f8068)