TRINITY-47 Normalised the favourites url for current user
https://sakaiproject.atlassian.net/browse/TRINITY-47

Seeing as favourites should only be pulled by and for the currently
logged in user, I've used the "current" token in the rest path to
emphasise that. Also, I removed the use of Resource in favour of
Autowired. We're trying to stick to the Spring annotations now, even
though it means we have to qualify with Qualifier.