SAK-47236 T&Q calculated questions grading: clear answerList map (#10493)
Although answerList should be empty on entering this function,
changes elsewhere have led to a case where it's not always empty.
This can lead to an answerMap that is sized larger than it should be,
depending on previous questions in the scoring process, leading to
item scores being divided by the wrong number of items for a question.

(cherry picked from commit 79bdb9527723591d4121e8d7ce0481159f776598)