Bring testing up to effectively full coverage
There are two branches uncovered: catch blocks for IO errors reading
directories after we checked that they were okay to read. Both are
logged with reasonable detail.

This also renamed TraditionalComponentTest to ComponentLoadingTest
because it is not strictly a class test. I still consider it a unit
test, but it's a functional unit with a considerable harness.