Rough in the loader/initializer structure
We need to add an initializer to be able to register the context once
it's created. We don't really have any other customization to do, but
the API doesn't quite let us catch it directly... so we have to
approximate the DelegatingSmartContextLoader. We lose Groovy support,
but we retain XML and Annotations, with a trivial implementation.

This still doesn't quite work, though, because the placeholder
properties are lost by switching to a GenericApplicationContext, which
is unfortunately required by the loaders with the real implementation.
But, registering those two beans up front should take care of it.