Add Sakai theme information to LTI launch data
Add the theme base color set in sakai properties for light and dark mode
as well as the user's dark mode preference to the LTI launch data. This
is primarily used in Tsugi to match the Tsugi theme and Tsugi tool theme
to whatever the theme is from the Sakai instance that launched the tool.