SAK-47475 grader: ensure docx opens in preview  panel
https://sakaiproject.atlassian.net/browse/SAK-47475

This also removes .odp and .odt from the file conversion list if you are
using the LibreOffice converter. viewerjs handles those types, so we
don't need the overhead on the server of converting them. .odp is also
stripped from the Apache Java converter in this change. Again, the
clientside viewer.js handles it.