Switch from SharedApplicationContext to GenericApplicationContext
Getting the property overrides and placeholders loading in the normal
startup phase was the last piece to using a standard Spring context. At
this point, the SharedApplicationContext is unused, and probably not
necessary. I'm leaving it in place for the moment, because it extends
AbstractRefreshableApplicationContext, which could theoretically open
the chance for reloadability.

However, if we can fix up SakaiProperties and use regular beans and
environment inheritance/resolution, we could load the kernel in a
generic context and let each component build its own child context. That
would require an entrypoint that delivers the parent context and
initiates component load, but that seems like the logical progression.

There doesn't seem to be much of a use case for kernel reloading if
components could be redeployed, especially if there is parallel or lazy
loading of tool webapps. The BDD/TDD cycle for tools can now be
dramatically faster and easier. Reloadable components (and probably
some proxy/reload trigger for the tools) would shorten the cycle time
for interactive testing and changes.