Support web and non-web context loading
The default SmartContextLoader behavior examines the configuration and
annotations on your test class. It selects one of four loaders, based on
whether or not you have @WebAppConfiguration and whether you have
classes or locations defined in the @ContextConfiguration. These are the
regular/web variations of annotation and XML loaders, respectively.

We want to preserve this behavior and not reimplement the details of
creating the context, but we have to plug in at exactly the right time:
we need to register context as global after it created, but before
refresh happens. Because loadContext is a template method, and it
delegates to the appropriate subclass for the load-refresh, we have to
extend all four variations to implement prepareContext, which is an
extension point for exactly this purpose -- preparing the context to be
refreshed.

The reason we need this timing is that the singletons are
"pre-instantiated" during refresh. Basically, any bean that gets an id
is a singleton, and all of our traditional services fall in this
category. Some of these call the ComponentManager cover during their
init method, so we need to have the global context registered already.
When it is, a ComponentManager.get translates to a context.getBean,
which is no problem; the container carries on creating any beans that
are requested, transitively.

All put together, this means that a test can inherit from the regular
test or the web test base class, depending on whether it needs a web
application context (and MockMvc / WebClient) facilities, and it will
have a full kernel and context set up, with all services injectable.