Move ToolPortal out to portal-util to make testing plausible
This isn't necessarily the right strategy, but it allows the portal to
continue to function the same and let us instantiate the
ToolPortalServlet in a test context. Combined with the RequestFilter,
which is already in kernel-api, we can drive very realistic requests to a
tool easily and see what else we need.