Resolve some class availability issues during tests
Because a different kind of Spring application context is now mentioned
in the ComponentManager cover, PowerMock has to deal with that. In these
few cases, it could not during tests, because spring-context wasn't
loaded. It's always provided in shared, anyway, so we just add it.

The other issue was that portal-tool now needs to know about the @Named
annotation, added so the ToolPortalServlet can be detected automatically
as a component/bean. A simple addition of javax.inject fixes that.