Move all of the test support classes to kernel-test
The recently added kernel-test and test-helper modules will probably
merge eventually, but for now we put all of the test support code in
kernel-test and pack it up in the helper jar like everything else. The
former gets a few new provided dependencies identified for compile.
This also adds test-helper to the kernel pom, so it is now within the
normal compile tree/cycle.

For now, anyway, the classes all have a Modi prefix. That's probably
unnecessary now that they're in a reasonable package.

This also turns down the log noise. If you want to turn logs up, put a
sakai.properties file in target/sakai-home of the module you are
testing with log.config entries.