SAK-47487 webapi add a client ping to detect when to trigger a cancel event #10679
https://sakaiproject.atlassian.net/browse/SAK-47487

Cancel events do not come upstream from clients unless they explicitly
close their SSE connection (js). So, if a session times out, or if the
tab is closed there is a possibility of the persistent HTTP connection
for that Sakai user hanging around indefinitely and exhausting Tomcat.
This change adds a ping SSE every x seconds (configurable from sakai
properties) which will cause the connection to be closed if the client
has disappeared.