SAK-43201 Kernel add a modern dependency injection (modi) ComponentManager (#10676)
* Begin work on modi -- Modern Dependency Injection

- Add Bootstrap listener so Tomcat loads components right away at startup
- Add concepts of a "traditional" component and a directory of them
- Set up basic component validation and exception

* Set up bean definition loading in TraditionalComponent

- Add Mockito and AssertJ for testing
- Build a ClassLoader with any classes and jars in the component
- Only load demo beans based on sakai.demo=true
- Load beans into the supplied registry

* Set up inverted bean loading

- Create context from Bootstrap
- Add BeanSource interface
- Register components on "starting"
- Test that a component can load and register its beans

* Move everything to component-manager module

- Rename BeanSource to BeanDefinitionSource
- Add the beginnings of the ComponentManagerShim
- Split Bootstrap into two pieces, adding Launcher
- Rename Bootstrap to TomcatListener
- Add LauncherTest to cover some of the startup w/fixtures

* Get things back to generally working... config-store circular dependency now

* Build out ComponentManagerShim

This adapter covers the entire interface of ComponentManager and a
couple of methods on SpringCompMgr because it is used directly in a few
places.

* Clean up and document the CM shim

* Build out Configuration loader

- Add Configuration class to handle the critical startup beans and
  customization (sakai-configuration.xml).

* Implement Overrides and demo mode

- Add ComponentOverrides to manage the convention, loading, and logging
- Cover Launcher with a bunch of tests for activation/override/demo

* Add documentation for BeanDefinitionSource and Launcher

* Reimplement and test sakai.home startup handling

* Extract and document SakaiHome and SysProp

These classes handle the setup of sakai.home and structured/discoverable
access to the core system properties.

* Rename TraditionalComponents to ComponentsDirectory

This is a little bit better name. This really does have to do with the
concrete directory, so it doesn't hurt to mention it here, and it
removes some overhead with having the singular and plural classes.

* Transform SakaiHome into Environment and build it out

- Move path calculation and enforcement into the Environment
- Handle catalina.base, sakai.home, sakai.components.root and
  sakai.security in the initialization with comprehensive tests

* Tidy up between the Launcher and Environment

- Roll back to one exception type
- Remove the Launcher constructor parameter in favor of the Environment
  all of that part.
- Clean up the straggling implementation in the Launcher

* Move the remaining conventions from Launcher into Environment

This is a nice, contained refactoring. It moves the conventions about
the main config file and overrides folder into Environment. It also
improves the ergonomics considerably, by having the Environment expose
all of these core settings as Paths.

* Update Javadoc and test class/jar loading for components

This introduces a class and jar within the component fixture structure
and tests that they are loaded properly. This brings the coverage
to very nearly 100%, with only an unlikely in-flight I/O problem untested
where a lib directory becomes unreadable after starting. That case is
handled with an initialization exception and detailed message.

The test class and jar are committed directly to avoid the complexity of
compiling and packaging during the test runs. We really only care that
we can load things, not build them.

* Test the Listener and Launcher fully for start/stop

* Bring testing up to effectively full coverage

There are two branches uncovered: catch blocks for IO errors reading
directories after we checked that they were okay to read. Both are
logged with reasonable detail.

This also renamed TraditionalComponentTest to ComponentLoadingTest
because it is not strictly a class test. I still consider it a unit
test, but it's a functional unit with a considerable harness.

* Proofread and run everything through autoformatting

* Allow MessagingService to do runtime registration

The requirement that handlers be set at initialization puts a dependency
outward from the kernel to other components. This allows the service to
start up empty and then let the handlers register when they come up.

* Switch bullhorn handlers over to explicit registration

* Change modi system property to sakai.modi.enabled

* SAK-43201 - Rename TomcatListener to SakaiStartListener