SAK-48105 rWiki search using incorrect namespace (#11148)
I'm speculating a bit, but I think rwiki was still using the old lucene style of searching?  I switched it so that the tool gets passed into the appropriate field instead of appended to the search terms.
Also, removed the html escaping from the results since we need to render the returned markup (matching search terms are bolded).  This should be consistent with the way the main search tool is working.