SAK-48697 UX grader fixes. JIT loading of submissions.
https://sakaiproject.atlassian.net/browse/SAK-48697

I've reduced the size of the json downloaded by ~5 times, and I've
removed an unnecessary fetch to load the grades. So, 5 times less
data, and half the http connections when grading. I've added a data
loading screen so the user knows what's going on. Also, I've added
paging so we can handle large class sizes. On the first load, only the
skeletons of the submissions are fetched, apart from the selected
submission, which is hydrated together with one either side of it. As
the user navigates using the next and previous buttons, the code looks
ahead and hydrates the next two unhydrated submissions.