SAK-48818 set due on defaultTask to be current epoch millis
https://sakaiproject.atlassian.net/browse/SAK-48818

The basic problem with this issue is that on FF change is only fired
when the time fields are populated, so due on the task is never set.
Setting the default due to the current epoch millis will populate the
datetime-local time fields and hopefully inform the user how to set
those fields.

Mozilla has an open ticket about actually showing the time picker
alongside the date picker, as Chrome does. *Hopefully*, this will be
coming down the pipeline soon.

https://bugzilla.mozilla.org/show_bug.cgi?id=1726108