SAK-50620 rubrics Prevent deletion of evaluated rubrics
https://sakaiproject.atlassian.net/browse/SAK-50620

A super user can now only delete a shared rubric, ie: one displayed in a
site other than the one it was originally authored in, if it does *not*
have any evaluations attached to it, so if it has never been used to
grade. This makes the shared rubric case consistent with the owner site
case where the delete button is never shown if the rubric has
evaluations.