SAK-50006 assignments Create task for students added after assignment
https://sakaiproject.atlassian.net/browse/SAK-50006

As part of this commit I also updated how "user.site.membership.add" is
fired. It was being fired with a semicolon separated set of key value
pairs. That data was not being used anywhere in the Sakai code base so I
replaced event.getResource with the user reference, which is the
"normal" way we do things. However, I do recall some conversation about
this string of data being used outside of Sakai in some event bus stuff.
Basically, we need a conversation about how to structure data in these
strings. I don't think it's a great idea to use semicolon as a
separator. Semicolons are valid characters to have in data values, imo.