SAK-51010 Calendar : frequency not working in Spanish with unset hour
The problem seems to come from a very old, very ugly js+vm related to the 24h vs am/pm formats, that should probably be rewritten entirely. In the meantime, I think it's fine to fix this specific bug by adding the same default value used by the English format.

The value is not even used, but it's just preventing from accessing and using the frequency feature.