SAK-51196 webcomponents Run tests on maven build
https://sakaiproject.atlassian.net/browse/SAK-51196

This change adds test scripts to the top level package.json that enable
us to test all the components at once using Playwright, and test just
the component which we are working on and its dependants. The full test
runs the tests in parallel and takes around 30s to run. The test:dev
script uses Lerna and will only test the staged files. The tool pom runs
the full component Playwright tests.