SAK-51224 conversations Create gb item when publishing topic
https://sakaiproject.atlassian.net/browse/SAK-51224

Rather than add a grading item when the topic is published, this PR
presents an error banner on the topic display if certain conditions are
present: topic.graded = true and topic.gradingItemId is undefined. The
user is presented with a banner stating that they need to edit the topic
and setup a grading item association, probably because the topic was
imported from an archive.

I've taken this approach because I would have needed to duplicate some
grading item data in the source topic and the xml archive to be able to
automagically create the same grading item. I didn't want to pollute the
topic entity just for that purpose.