Update test-helper dependencies for Sakai 25+
- json-simple
- javax-mail
- ehcache-core

We add json-simple because HtmlPageFilter gained this dependency (which
is widely used outside of the kernel) in support of configuring MathJax.
These should probably be converted across the board to Jackson or an
alternative implementation like jolokia-json.

We add com.sun.mail:jakarta-mail to support changes in the Message
Service implementation. Whether we stay with that jar or switch to
something like Angus, the references should all be updated to jakarta.

The ehcache-core dependency is dubious, with both Ignite and Hazelcast
in place.